+++ Common/DtaDevOSDrive.h
@@ -72,7 +72,7 @@ public:
    * @param buffer input/output buffer
    * @param bufferlen length of the input/output buffer
    */
-  virtual int sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
+  virtual uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
                       void * buffer, unsigned int bufferlen) = 0;
 
   /** Routine to identify the device and fill out the device info structure.
