+++ macos/DtaDevMacOSSata.h
@@ -35,7 +35,7 @@ public:
    * @param buffer input/output buffer
    * @param bufferlen length of the input/output buffer
    */
-  virtual int sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
+  virtual uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
                           void * buffer, uint32_t bufferlen);
